Feat: Implement the playground update feature

+++ src/main/java/com/bookmysport/authentication_service/Models/ServiceProviderModel.java
@@ -38,7 +38,7 @@ public class ServiceProviderModel {
 
     @NotNull
     @NumberFormat(pattern = "Enter a number")
-    private int phoneNumber;
+    private long phoneNumber;
 
     @NotNull
     private String password;
