+++ src/main/java/com/bookmysport/authentication_service/Models/UserModel.java
@@ -38,7 +38,7 @@ public class UserModel {
 
     @NotNull
     @NumberFormat(pattern = "Enter a number")
-    private int phoneNumber;
+    private long phoneNumber;
 
     @NotNull
     private String password;
